Various fixes / improvements [skip ci]

+++ scripts/helpers/eosio.bash
@@ -38,12 +38,17 @@ function setup() {
         echo "PIN_COMPILER: ${PIN_COMPILER}"
         echo "BUILD_CLANG: ${BUILD_CLANG}"
     fi
-    [[ -d ./build ]] && execute rm -rf ./build # cleanup old build directory
+    [[ -d $BUILD_DIR ]] && execute rm -rf $BUILD_DIR # cleanup old build directory
     # Be absolutely sure TEMP_DIR is set (we could do damage here)
     ## IF user set, they can be responsible for creating and then removing stale data
-    if [[ $TEMP_DIR =~ "/tmp" ]]; then 
-        execute mkdir -p $TEMP_DIR
-        execute rm -rf $TEMP_DIR/*
+    if [[ $TEMP_DIR =~ "/tmp" ]]; then
+        if [[ ! -d $TEMP_DIR ]]; then
+            DRYRUN_ORIGINAL=$DRYRUN
+            DRYRUN=false
+            execute mkdir -p $TEMP_DIR
+            execute rm -rf $TEMP_DIR/*
+            DRYRUN=$DRYRUN_ORIGINAL
+        fi
     fi
     execute mkdir -p $BUILD_DIR
     execute mkdir -p $SRC_LOCATION
@@ -136,10 +141,10 @@ function ensure-compiler() {
 }
 
 function ensure-cmake() {
-    echo "${COLOR_CYAN}[Checking CMAKE installation]${COLOR_NC}"
+    echo "${COLOR_CYAN}[Ensuring CMAKE installation]${COLOR_NC}"
     if [[ ! -e "${CMAKE}" ]]; then
-        echo "Installing CMAKE..."
-        execute bash -c "curl -LO https://cmake.org/files/v${CMAKE_VERSION_MAJOR}.${CMAKE_VERSION_MINOR}/cmake-${CMAKE_VERSION}.tar.gz \
+		execute bash -c "cd $SRC_LOCATION && \
+        curl -LO https://cmake.org/files/v${CMAKE_VERSION_MAJOR}.${CMAKE_VERSION_MINOR}/cmake-${CMAKE_VERSION}.tar.gz \
         && tar -xzf cmake-${CMAKE_VERSION}.tar.gz \
         && cd cmake-${CMAKE_VERSION} \
         && ./bootstrap --prefix=${EOSIO_HOME} \
@@ -149,36 +154,42 @@ function ensure-cmake() {
         && rm -f cmake-${CMAKE_VERSION}.tar.gz"
         [[ -z "${CMAKE}" ]] && export CMAKE="${BIN_LOCATION}/cmake"
         echo " - CMAKE successfully installed @ ${CMAKE}"
+        echo ""
     else
         echo " - CMAKE found @ ${CMAKE}."
+        echo ""
     fi
 }
 
 function ensure-zlib() {
     if $PIN_COMPILER; then
-        echo "${COLOR_CYAN}[Checking zlib library installation]${COLOR_NC}"
+        echo "${COLOR_CYAN}[Ensuring zlib library installation]${COLOR_NC}"
         if [[ ! -d $OPT_LOCATION/zlib ]]; then
-            echo "Installing zlib..."
-            execute bash -c "curl -LO https://www.zlib.net/zlib-${ZLIB_VERSION}.tar.gz && tar -xf zlib-${ZLIB_VERSION}.tar.gz \
+		    execute bash -c "cd $SRC_LOCATION && \
+            curl -LO https://www.zlib.net/zlib-${ZLIB_VERSION}.tar.gz && tar -xf zlib-${ZLIB_VERSION}.tar.gz \
             && cd zlib-${ZLIB_VERSION} && mkdir build && cd build \
             && ../configure --prefix=$OPT_LOCATION/zlib \
             && make -j$JOBS install"
+            echo " - zlib library successfully installed @ $OPT_LOCATION/zlib"
+            echo ""
+        else
+            echo " - zlib library found @ $OPT_LOCATION/zlib"
+            echo ""
         fi
     fi
 }
 
 function ensure-boost() {
     [[ $ARCH == "Darwin" ]] && export CPATH="$(python-config --includes | awk '{print $1}' | cut -dI -f2):$CPATH" # Boost has trouble finding pyconfig.h
-    echo "${COLOR_CYAN}[Checking Boost $( echo $BOOST_VERSION | sed 's/_/./g' ) library installation]${COLOR_NC}"
+    echo "${COLOR_CYAN}[Ensuring Boost $( echo $BOOST_VERSION | sed 's/_/./g' ) library installation]${COLOR_NC}"
     BOOSTVERSION=$( grep "#define BOOST_VERSION" "$EOSIO_HOME/opt/boost/include/boost/version.hpp" 2>/dev/null | tail -1 | tr -s ' ' | cut -d\  -f3 || true )
     if [[ "${BOOSTVERSION}" != "${BOOST_VERSION_MAJOR}0${BOOST_VERSION_MINOR}0${BOOST_VERSION_PATCH}" ]]; then
+        B2_FLAGS="-q -j${JOBS} --with-iostreams --with-date_time --with-filesystem --with-system --with-program_options --with-chrono --with-test install"
         if $PIN_COMPILER; then
-            B2_FLAGS="toolset=clang cxxflags=\"-stdlib=libc++ -D__STRICT_ANSI__ -nostdinc++ -I${CLANG_ROOT}/include/c++/v1\" linkflags=\"-stdlib=libc++\" link=static threading=multi --with-iostreams --with-date_time --with-filesystem --with-system --with-program_options --with-chrono --with-test -q -j${JOBS} -sZLIB_LIBRARY_PATH=\"${OPT_LOCATION}/zlib/lib\" -sZLIB_INCLUDE=\"${OPT_LOCATION}/zlib/include\" -sZLIB_SOURCE=\"${SRC_LOCATION}/zlib-${ZLIB_VERSION}\" install"
-        else
-            B2_FLAGS="-q -j${JOBS} --with-iostreams --with-date_time --with-filesystem --with-system --with-program_options --with-chrono --with-test install"
-        fi 
-        echo "Installing Boost library..."
-        execute bash -c "curl -LO https://dl.bintray.com/boostorg/release/$BOOST_VERSION_MAJOR.$BOOST_VERSION_MINOR.$BOOST_VERSION_PATCH/source/boost_$BOOST_VERSION.tar.bz2 \
+            B2_FLAGS="toolset=clang cxxflags='-stdlib=libc++ -D__STRICT_ANSI__ -nostdinc++ -I${CLANG_ROOT}/include/c++/v1' linkflags='-stdlib=libc++' link=static threading=multi --with-iostreams --with-date_time --with-filesystem --with-system --with-program_options --with-chrono --with-test -q -j${JOBS} -sZLIB_LIBRARY_PATH=\"${OPT_LOCATION}/zlib/lib\" -sZLIB_INCLUDE=\"${OPT_LOCATION}/zlib/include\" -sZLIB_SOURCE=\"${SRC_LOCATION}/zlib-${ZLIB_VERSION}\" install"
+        fi
+		execute bash -c "cd $SRC_LOCATION && \
+        curl -LO https://dl.bintray.com/boostorg/release/$BOOST_VERSION_MAJOR.$BOOST_VERSION_MINOR.$BOOST_VERSION_PATCH/source/boost_$BOOST_VERSION.tar.bz2 \
         && tar -xjf boost_$BOOST_VERSION.tar.bz2 \
         && cd $BOOST_ROOT \
         && ./bootstrap.sh --prefix=$BOOST_ROOT \
@@ -187,21 +198,22 @@ function ensure-boost() {
         && rm -f boost_$BOOST_VERSION.tar.bz2 \
         && rm -rf $BOOST_LINK_LOCATION \
         && ln -s $BOOST_ROOT $BOOST_LINK_LOCATION"
-        echo " - Boost library successfully installed @ ${BOOST_ROOT}."
+        echo " - Boost library successfully installed @ ${BOOST_ROOT}"
+        echo ""
     else
-        echo " - Boost library found with correct version @ ${BOOST_ROOT}."
+        echo " - Boost library found with correct version @ ${BOOST_ROOT}"
+        echo ""
     fi
 }
 
 function ensure-llvm() {
-    echo "${COLOR_CYAN}[Checking LLVM 4 support]${COLOR_NC}"
+    echo "${COLOR_CYAN}[Ensuring LLVM 4 support]${COLOR_NC}"
     if [[ ! -d $LLVM_ROOT ]]; then
-        echo "Installing LLVM 4..."
         if [[ $ARCH == "Darwin" ]]; then # Handle brew installed llvm@4
 	        execute ln -s /usr/local/opt/llvm@4 $LLVM_ROOT
 	        echo " - LLVM successfully linked from /usr/local/opt/llvm@4 to ${LLVM_ROOT}"
         else
-            if $PIN_COMPILER; then
+            if $PIN_COMPILER || $BUILD_CLANG; then
                 CMAKE_FLAGS="-DCMAKE_INSTALL_PREFIX='${LLVM_ROOT}' -DLLVM_TARGETS_TO_BUILD=host -DLLVM_BUILD_TOOLS=false -DLLVM_ENABLE_RTTI=1 -DCMAKE_BUILD_TYPE=Release -DCMAKE_TOOLCHAIN_FILE='${BUILD_DIR}/pinned_toolchain.cmake' .."
             else
                 if [[ $NAME == "Ubuntu" ]]; then
@@ -211,27 +223,27 @@ function ensure-llvm() {
                 fi
                 CMAKE_FLAGS="-G 'Unix Makefiles' -DCMAKE_INSTALL_PREFIX=${LLVM_ROOT} -DLLVM_TARGETS_TO_BUILD='host' -DLLVM_BUILD_TOOLS=false -DLLVM_ENABLE_RTTI=1 -DCMAKE_BUILD_TYPE=Release .."
             fi
-            execute bash -c "cd ../opt \
+            execute bash -c "cd ${OPT_LOCATION} \
             && git clone --depth 1 --single-branch --branch $LLVM_VERSION https://github.com/llvm-mirror/llvm.git llvm && cd llvm \
             && mkdir build \
             && cd build \
             && ${CMAKE} ${CMAKE_FLAGS} \
             && make -j${JOBS} \
-            && make install \
-            && cd ../.."
+            && make install"
             echo " - LLVM successfully installed @ ${LLVM_ROOT}"
+            echo ""
         fi
     else
         echo " - LLVM found @ ${LLVM_ROOT}."
+        echo ""
     fi
 }
 
 
 function build-clang() {
     if $BUILD_CLANG; then
-        echo "${COLOR_CYAN}[Checking Clang support]${COLOR_NC}"
+        echo "${COLOR_CYAN}[Ensuring Clang support]${COLOR_NC}"
         if [[ ! -d $CLANG_ROOT ]]; then
-            echo "Installing Clang 8..."
             execute bash -c "cd ${TEMP_DIR} \
             && git clone --single-branch --branch $PINNED_COMPILER_BRANCH https://git.llvm.org/git/llvm.git clang8 \
             && cd clang8 && git checkout $PINNED_COMPILER_LLVM_COMMIT \
@@ -256,13 +268,15 @@ function build-clang() {
             && cd compiler-rt && git checkout $PINNED_COMPILER_COMPILER_RT_COMMIT && cd ../ \
             && cd ${TEMP_DIR}/clang8 \
             && mkdir build && cd build \
-            && ${CMAKE} -G \"Unix Makefiles\" -DCMAKE_INSTALL_PREFIX=\"${CLANG_ROOT}\" -DLLVM_BUILD_EXTERNAL_COMPILER_RT=ON -DLLVM_BUILD_LLVM_DYLIB=ON -DLLVM_ENABLE_LIBCXX=ON -DLLVM_ENABLE_RTTI=ON -DLLVM_INCLUDE_DOCS=OFF -DLLVM_OPTIMIZED_TABLEGEN=ON -DLLVM_TARGETS_TO_BUILD=all -DCMAKE_BUILD_TYPE=Release .. \
+            && ${CMAKE} -G 'Unix Makefiles' -DCMAKE_INSTALL_PREFIX='${CLANG_ROOT}' -DLLVM_BUILD_EXTERNAL_COMPILER_RT=ON -DLLVM_BUILD_LLVM_DYLIB=ON -DLLVM_ENABLE_LIBCXX=ON -DLLVM_ENABLE_RTTI=ON -DLLVM_INCLUDE_DOCS=OFF -DLLVM_OPTIMIZED_TABLEGEN=ON -DLLVM_TARGETS_TO_BUILD=all -DCMAKE_BUILD_TYPE=Release .. \
             && make -j${JOBS} \
             && make install \
-            && cd ../.."
-            echo " - Clang 8 successfully installed @ ${CLANG_ROOT}."
+            && rm -rf ${TEMP_DIR}/*"
+            echo " - Clang 8 successfully installed @ ${CLANG_ROOT}"
+            echo ""
         else
-            echo " - Clang 8 found @ ${CLANG_ROOT}."
+            echo " - Clang 8 found @ ${CLANG_ROOT}"
+            echo ""
         fi
         export CXX=$CPP_COMP
         export CC=$CC_COMP
@@ -279,7 +293,7 @@ function ensure-yum-packages() {
         _2=("$(echo $2 | sed 's/-s /-s\n/g')")
         for ((i = 0; i < ${#_2[@]}; i++)); do echo "${_2[$i]}\n" | sed 's/-s\\n/-s/g' >> $DEPS_FILE; done
     fi
-    echo "${COLOR_CYAN}[Checking YUM installation]${COLOR_NC}"
+    echo "${COLOR_CYAN}[Ensuring YUM installation]${COLOR_NC}"
     if ! YUM=$( command -v yum 2>/dev/null ); then echo " - YUM must be installed to compile EOS.IO." && exit 1
     else echo "Yum installation found at ${YUM}."; fi
     while true; do
@@ -299,7 +313,7 @@ function ensure-yum-packages() {
         esac
     done
     if [[ $NAME == "CentOS Linux" ]]; then
-        echo "${COLOR_CYAN}[Checking installation of Centos Software Collections Repository]${COLOR_NC}" # Needed for rh-python36
+        echo "${COLOR_CYAN}[Ensuring installation of Centos Software Collections Repository]${COLOR_NC}" # Needed for rh-python36
         SCL=$( rpm -qa | grep -E 'centos-release-scl-[0-9].*' || true )
         if [[ -z "${SCL}" ]]; then
             while true; do
@@ -322,7 +336,7 @@ function ensure-yum-packages() {
             echo " - ${SCL} found."
         fi
     fi
-    echo "${COLOR_CYAN}[Checking for installed package dependencies]${COLOR_NC}"
+    echo "${COLOR_CYAN}[Ensuring for installed package dependencies]${COLOR_NC}"
     OLDIFS="$IFS"; IFS=$','
     while read -r testee tester; do
         if [[ ! -z $(eval $tester $testee) ]]; then
@@ -362,7 +376,7 @@ function ensure-brew-packages() {
         _2=("$(echo $2 | sed 's/-s /-s\n/g')")
         for ((i = 0; i < ${#_2[@]}; i++)); do echo "${_2[$i]}\n" | sed 's/-s\\n/-s/g' >> $DEPS_FILE; done
     fi
-    echo "${COLOR_CYAN}[Checking HomeBrew dependencies]${COLOR_NC}"
+    echo "${COLOR_CYAN}[Ensuring HomeBrew dependencies]${COLOR_NC}"
     OLDIFS="$IFS"
     IFS=$','
     while read -r name path; do
@@ -418,13 +432,12 @@ function ensure-apt-packages() {
     if [[ -n "${2}" ]]; then # Handle EXTRA_DEPS passed in and add them to temp DEPS_FILE
         OLDIFS="$IFS"; IFS=$''
         _2=("$(echo $2 | sed 's/-s /-s\n/g')")
-        for ((i = 0; i < ${#_2[@]}; i++)); do echo "${_2[$i]}\n" | sed 's/-s\\n/-s/g' >> $DEPS_FILE; done
+        for ((i = 0; i < ${#_2[@]}; i++)); do echo "${_2[$i]}" >> $DEPS_FILE; done
     fi
-    echo "${COLOR_CYAN}[Checking for installed package dependencies]${COLOR_NC}"
+    echo "${COLOR_CYAN}[Ensuring for installed package dependencies]${COLOR_NC}"
     OLDIFS="$IFS"; IFS=$','
-
     while read -r testee tester; do
-        if [[ ! -z $(eval $tester $testee) ]]; then
+        if [[ ! -z $(eval $tester $testee 2>/dev/null) ]]; then
             echo " - ${testee} ${COLOR_GREEN}found!${COLOR_NC}"
         else
             DEPS=$DEPS"${testee} "
@@ -432,9 +445,9 @@ function ensure-apt-packages() {
             (( COUNT++ ))
         fi
     done < $DEPS_FILE
-    echo ""
     IFS=$OLDIFS
     if [[ $COUNT > 0 ]]; then
+        echo ""
         while true; do
             [[ $NONINTERACTIVE == false ]] && read -p "${COLOR_YELLOW}Do you wish to install missing dependencies? (y/n)?${COLOR_NC} " PROCEED
             case $PROCEED in
@@ -448,5 +461,6 @@ function ensure-apt-packages() {
         done
     else
         echo " - No required dependencies to install."
+        echo ""
     fi
 }
